Move state machine mixin tests into a module
